minimal AMD family history
also fix returning a feature as in a family if the family has zero cpus.
the number of cpus with the feature was the same as the number of cpus
in the family so all cpus in the family had the feature. but if there
are zero cpus in family then none have the feature, and this case was
missing.